feat: add the new account fraud use case demo (#206)
Also includes a minor fix to the Button component to pass the `disabled`
property through to the `disabled` attribute on the `button` element.
This allows the `toBeDisabled` assertion to be used within a Playwright
test.

* fix: add missing period in 1st instruction step

Co-authored-by: Juraj Uhlar <juro.uhlar@gmail.com>

* style: improve readability of create-account API implementation

* fix: form should not be disabled after a failure to create an additional trail

* style: temporarily override title for account creation fraud

In lieu of the changes for #204, this commit does not use the long SEO
title on the home page to avoid an unnecessary distraction during the
design review.

* style: add icon for account creation fraud demo

* fix: add new account-creation-fraud icon that fixes style issues

* docs: add link to case study for account creation fraud demo

* style: update content to use new titleH1 property

* style: rename account-creation-fraud demo to new-account-fraud demo

The use case tutorial is getting a refreshed name. Align the naming
within the code and the content with the new name.

---------

Co-authored-by: Juraj Uhlar <juro.uhlar@gmail.com>